feat(#31): implement DepositRepository with derived query methods for quote and payment reference, also renaming package from model to entity
