+++ src/main/java/com/remittance/deposit/entity/Deposit.java
@@ -1,7 +1,7 @@
-package com.remittance.deposit.model;
+package com.remittance.deposit.entity;
 
 import com.remittance.common.model.BaseEntity;
-import com.remittance.quote.model.Quote;
+import com.remittance.quote.entity.Quote;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
