+++ src/main/java/com/remittance/deposit/entity/DepositStatus.java
@@ -1,4 +1,4 @@
-package com.remittance.deposit.model;
+package com.remittance.deposit.entity;
 
 public enum DepositStatus {
     PENDING,
